+++ rules/inclusive-gateway.js
@@ -1,6 +1,6 @@
 const { is } = require('bpmnlint-utils');
 
-const { hasProperties, ERROR_TYPES } = require('./utils/element');
+const { ERROR_TYPES } = require('./utils/element');
 
 const { reportErrors } = require('./utils/reporter');
 
@@ -25,24 +25,6 @@ module.exports = function() {
       };
 
       reportErrors(node, reporter, error);
-    }
-
-    const outgoing = node.get('outgoing');
-
-    if (outgoing && outgoing.length > 1) {
-      for (let sequenceFlow of outgoing) {
-        if (node.get('default') !== sequenceFlow) {
-          const errors = hasProperties(sequenceFlow, {
-            conditionExpression: {
-              required: true
-            }
-          }, sequenceFlow);
-
-          if (errors.length) {
-            reportErrors(sequenceFlow, reporter, errors);
-          }
-        }
-      }
     }
   }
