+++ test/camunda-cloud/inclusive-gateway.spec.js
@@ -11,55 +11,15 @@ const { ERROR_TYPES } = require('../../rules/utils/element');
 
 const valid = [
   {
-    name: 'inclusive gateway (1 outgoing sequence flow)',
+    name: 'inclusive gateway (1 incoming sequence flow)',
     moddleElement: createModdle(createProcess(`
+      <bpmn:startEvent id="StartEvent_1">
+        <bpmn:outgoing>SequenceFlow_1</bpmn:outgoing>
+      </bpmn:startEvent>
       <bpmn:inclusiveGateway id="InclusiveGateway_1">
-        <bpmn:outgoing>SequenceFlow_1</bpmn:outgoing>
-      </bpmn:inclusiveGateway>
-      <bpmn:endEvent id="EndEvent_1">
         <bpmn:incoming>SequenceFlow_1</bpmn:incoming>
-      </bpmn:endEvent>
-      <bpmn:sequenceFlow id="SequenceFlow_1" sourceRef="InclusiveGateway_1" targetRef="EndEvent_1" />
-    `))
-  },
-  {
-    name: 'inclusive gateway (2 outgoing sequence flows with condition)',
-    moddleElement: createModdle(createProcess(`
-      <bpmn:inclusiveGateway id="InclusiveGateway_1">
-        <bpmn:outgoing>SequenceFlow_1</bpmn:outgoing>
-        <bpmn:outgoing>SequenceFlow_2</bpmn:outgoing>
       </bpmn:inclusiveGateway>
-      <bpmn:endEvent id="EndEvent_1">
-        <bpmn:incoming>SequenceFlow_1</bpmn:incoming>
-      </bpmn:endEvent>
-      <bpmn:sequenceFlow id="SequenceFlow_1" sourceRef="InclusiveGateway_1" targetRef="EndEvent_1">
-        <bpmn:conditionExpression xsi:type="bpmn:tFormalExpression">=foo</bpmn:conditionExpression>
-      </bpmn:sequenceFlow>
-      <bpmn:endEvent id="EndEvent_2">
-        <bpmn:incoming>SequenceFlow_2</bpmn:incoming>
-      </bpmn:endEvent>
-      <bpmn:sequenceFlow id="SequenceFlow_2" sourceRef="InclusiveGateway_1" targetRef="EndEvent_2">
-        <bpmn:conditionExpression xsi:type="bpmn:tFormalExpression">=bar</bpmn:conditionExpression>
-      </bpmn:sequenceFlow>
-    `))
-  },
-  {
-    name: 'inclusive gateway (2 outgoing sequence flows, 1 default, 1 with condition)',
-    moddleElement: createModdle(createProcess(`
-      <bpmn:inclusiveGateway id="InclusiveGateway_1" default="SequenceFlow_2">
-        <bpmn:outgoing>SequenceFlow_1</bpmn:outgoing>
-        <bpmn:outgoing>SequenceFlow_2</bpmn:outgoing>
-      </bpmn:inclusiveGateway>
-      <bpmn:endEvent id="EndEvent_1">
-        <bpmn:incoming>SequenceFlow_1</bpmn:incoming>
-      </bpmn:endEvent>
-      <bpmn:sequenceFlow id="SequenceFlow_1" sourceRef="InclusiveGateway_1" targetRef="EndEvent_1">
-        <bpmn:conditionExpression xsi:type="bpmn:tFormalExpression">=foo</bpmn:conditionExpression>
-      </bpmn:sequenceFlow>
-      <bpmn:endEvent id="EndEvent_2">
-        <bpmn:incoming>SequenceFlow_2</bpmn:incoming>
-      </bpmn:endEvent>
-      <bpmn:sequenceFlow id="SequenceFlow_2" sourceRef="InclusiveGateway_1" targetRef="EndEvent_2" />
+      <bpmn:sequenceFlow id="SequenceFlow_1" sourceRef="StartEvent_1" targetRef="InclusiveGateway_1" />
     `))
   },
   {
@@ -102,83 +62,6 @@ const invalid = [
         property: 'incoming'
       }
     }
-  },
-  {
-    name: 'inclusive gateway (2 outgoing sequence flows, one without condition)',
-    moddleElement: createModdle(createProcess(`
-      <bpmn:inclusiveGateway id="InclusiveGateway_1">
-        <bpmn:outgoing>SequenceFlow_1</bpmn:outgoing>
-        <bpmn:outgoing>SequenceFlow_2</bpmn:outgoing>
-      </bpmn:inclusiveGateway>
-      <bpmn:endEvent id="EndEvent_1">
-        <bpmn:incoming>SequenceFlow_1</bpmn:incoming>
-      </bpmn:endEvent>
-      <bpmn:sequenceFlow id="SequenceFlow_1" sourceRef="InclusiveGateway_1" targetRef="EndEvent_1">
-        <bpmn:conditionExpression xsi:type="bpmn:tFormalExpression">=foo</bpmn:conditionExpression>
-      </bpmn:sequenceFlow>
-      <bpmn:endEvent id="EndEvent_2">
-        <bpmn:incoming>SequenceFlow_2</bpmn:incoming>
-      </bpmn:endEvent>
-      <bpmn:sequenceFlow id="SequenceFlow_2" sourceRef="InclusiveGateway_1" targetRef="EndEvent_2" />
-    `)),
-    report: {
-      id: 'SequenceFlow_2',
-      message: 'Element of type <bpmn:SequenceFlow> must have property <conditionExpression>',
-      path: [
-        'conditionExpression'
-      ],
-      data: {
-        type: ERROR_TYPES.PROPERTY_REQUIRED,
-        node: 'SequenceFlow_2',
-        parentNode: null,
-        requiredProperty: 'conditionExpression'
-      }
-    }
-  },
-  {
-    name: 'inclusive gateway (2 outgoing sequence flows, 2 without condition)',
-    moddleElement: createModdle(createProcess(`
-      <bpmn:inclusiveGateway id="InclusiveGateway_1">
-        <bpmn:outgoing>SequenceFlow_1</bpmn:outgoing>
-        <bpmn:outgoing>SequenceFlow_2</bpmn:outgoing>
-      </bpmn:inclusiveGateway>
-      <bpmn:endEvent id="EndEvent_1">
-        <bpmn:incoming>SequenceFlow_1</bpmn:incoming>
-      </bpmn:endEvent>
-      <bpmn:sequenceFlow id="SequenceFlow_1" sourceRef="InclusiveGateway_1" targetRef="EndEvent_1" />
-      <bpmn:endEvent id="EndEvent_2">
-        <bpmn:incoming>SequenceFlow_2</bpmn:incoming>
-      </bpmn:endEvent>
-      <bpmn:sequenceFlow id="SequenceFlow_2" sourceRef="InclusiveGateway_1" targetRef="EndEvent_2" />
-    `)),
-    report: [
-      {
-        id: 'SequenceFlow_1',
-        message: 'Element of type <bpmn:SequenceFlow> must have property <conditionExpression>',
-        path: [
-          'conditionExpression'
-        ],
-        data: {
-          type: ERROR_TYPES.PROPERTY_REQUIRED,
-          node: 'SequenceFlow_1',
-          parentNode: null,
-          requiredProperty: 'conditionExpression'
-        }
-      },
-      {
-        id: 'SequenceFlow_2',
-        message: 'Element of type <bpmn:SequenceFlow> must have property <conditionExpression>',
-        path: [
-          'conditionExpression'
-        ],
-        data: {
-          type: ERROR_TYPES.PROPERTY_REQUIRED,
-          node: 'SequenceFlow_2',
-          parentNode: null,
-          requiredProperty: 'conditionExpression'
-        }
-      }
-    ]
   }
 ];
